Fix the wallpaper not hiding immediately
When the application switches to the foreground, the wallpaper should
immediately hide after the shell transition ends. However, the current
issue arises because we are collecting the wallpaper target as
WindowState instead of its corresponding token, which causes the hiding
operation to be delayed.

To address this, we can add some checks: if the target is WindowState,
we will attempt to convert its token to WallpaperWindowToken in order
to resolve this issue.

Bug: 381804828
Test: Manual
Flag: EXEMPT bugfix

Google: 3385960
Change-Id: I5679a759811b7acb44909d840579fd7fb291412d